backend/requirements.txt
django
django-cors-headers
django-filter
django-rest-knox
djangorestframework
djangorestframework-simplejwt[crypto]
drf-spectacular
gunicorn
psycopg2-binary
uvicorn
